Move category code into `pcats`.
