+++ src/gpsea/analysis/__init__.py
@@ -1,4 +1,5 @@
 from ._api import CohortAnalysis, GenotypePhenotypeAnalysisResult, HpoMtcReport
+# TODO This should go away
 from ._config import CohortAnalysisConfiguration, configure_cohort_analysis, configure_default_protein_metadata_service, MtcStrategy
 from ._gp_analysis import apply_predicates_on_patients
 from ._util import prepare_hpo_terms_of_interest
