+++ src/gpsea/analysis/pcats/__init__.py
@@ -1,5 +1,5 @@
 """
-The `gpsea.analysis.multip` tests the association between genotype and phenotype groups,
+The `gpsea.analysis.pcats` tests the association between genotype and phenotype groups,
 if the groups can be defined in terms of discrete, unique, and non-overlapping categories.
 
 Each individual is assigned into a genotype and phenotype group
@@ -7,8 +7,8 @@ using :class:`~gpsea.analysis.predicate.genotype.GenotypePolyPredicate`
 and :class:`~gpsea.analysis.predicate.phenotype.PhenotypePolyPredicate` respectively.
 
 A contingency matrix with group counts is prepared
-and the counts are tested for association using :class:`~gpsea.analysis.stats.CountStatistic`,
-such as :class:`~gpsea.analysis.stats.ScipyFisherExact`.
+and the counts are tested for association using :class:`~gpsea.analysis.pcats.stats.CountStatistic`,
+such as :class:`~gpsea.analysis.pcats.stats.ScipyFisherExact`.
 
 It is typical to test several phenotype groups at the same time.
 Therefore, we must correct for multiple testing to prevent false positive findings.
