+++ src/gpsea/analysis/pcats/stats/_test__stats.py
@@ -1,5 +1,3 @@
-from contextlib import nullcontext as does_not_raise
-
 import numpy as np
 import pandas as pd
 import pytest
