+++ src/gpsea/analysis/pscore/__init__.py
@@ -1,0 +1,5 @@
+from ._impl import CountingPhenotypeScorer
+
+__all__ = [
+    'CountingPhenotypeScorer',
+]
