+++ src/gpsea/preprocessing/_config.py
@@ -359,7 +359,7 @@ def load_phenopackets(
     """
     Map phenopacket JSON file into patient, validate the patient data, and assemble the patients into a cohort.
 
-    :param pp_directory: path to a folder with phenopacket JSON files. An error is raised if the path does not point to
+    :param phenopackets: path to a folder with phenopacket JSON files. An error is raised if the path does not point to
       a directory with at least one phenopacket.
     :param cohort_creator: cohort creator for turning a sequence of phenopacket
       into a :class:`~gpsea.model.Cohort`.
